+++ door/validator.go
@@ -12,7 +12,7 @@ type (
 	}
 
 	ContractInfoer interface {
-		IsAllowedAt(booking BookingID, renter RenterPublicKey, timestamp int)  (bool, error)
+		IsAllowedAt(booking BookingID, renter RenterPublicKey, timestamp int64)  (bool, error)
 	}
 	Lock interface {
 		Open()
@@ -21,7 +21,7 @@ type (
 
 func (v *Validator) Handle(m OpenDoorMessage) error {
 	now := time.Now()
-	ok, err := v.ContractInfoer.IsAllowedAt(m.Booking, m.RenterPubkey, int(now.Unix()))
+	ok, err := v.ContractInfoer.IsAllowedAt(m.Booking, m.RenterPubkey, now.Unix())
 	if err != nil {
 		log.Println("Failed to confirm contract.")
 		return err
